SharedPreferences backed action lookup
How this is working now;

We populate a SharedPreferences file with String elements.

The key's for the strings are the "MCU KEYCODE" run through
Integer.toString(..).

The values are made up of;
first character: a number from 1-3 defined in Constants.class that
identifies the type of action to be taken -- start activity, send
broadcast, inject keycode.
second character onward: the actual value.